pyproject.toml
[tool.poetry]
authors = ["Francisco Del Roio <francipvb@hotmail.com>"]
description = ""
name = "socketio-chat"
version = "0.1.0"

[tool.poetry.dependencies]
SQLAlchemy = {extras = ["postgresql_asyncpg", "aiosqlite"], version = "^1.4.35"}
aiofiles = "^0.8.0"
fastapi = "^0.75.1"
httptools = "^0.4.0"
pydantic = {extras = ["emails", "dotenv"], version = "^1.9.0"}
python = "^3.10"
python-socketio = "^5.5.2"
uvicorn = "^0.17.6"
wsproto = "^1.1.0"
orjson = "^3.6.7"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.3.0"
flake8 = "^4.0.1"
ipdb = "^0.13.9"
isort = "^5.10.1"
mypy = "^0.942"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
devtools = "^0.8.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
